ocr/paddlex-layout/pyproject.toml
[project]
name = "paddlex-layout-liteparse"
version = "0.1.0"
description = "PaddleX layout detection and formula recognition server for LiteParse"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi>=0.131.0",
    "numpy>=2.4.2",
    "paddlex[ocr]>=3.0.0",
    "paddlepaddle==3.0.0",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]

[tool.uv.sources]
paddlepaddle = { index = "paddle" }

[[tool.uv.index]]
name = "paddle"
url = "https://www.paddlepaddle.org.cn/packages/stable/cpu/"
explicit = true

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
